Fixed location for service function (relocated to GenreService) and added controller function for getTop5Genres
